Move tealium stuff and cleanup
