+++ services/app-web/src/components/FileUpload/FileItemList/FileItemsList.test.tsx
@@ -4,13 +4,8 @@ import userEvent from '@testing-library/user-event'
 
 import { FileItemT } from '../FileProcessor/FileProcessor'
 import { FileItemsList } from './FileItemsList'
-import { TEST_PDF_FILE } from '../../../testHelpers/jestHelpers'
+import { TEST_PDF_FILE } from '../../../testHelpers'
 import * as tealium from '../../../hooks/useTealium'
-import {
-    TealiumButtonEventObject,
-    TealiumInternalLinkEventObject,
-    TealiumEventObjectTypes,
-} from '../../../constants/tealium'
 
 describe('FileItemList component', () => {
     const pending: FileItemT = {
@@ -73,20 +68,10 @@ describe('FileItemList component', () => {
 
     beforeEach(() => {
         spyOnUseTealium.mockImplementation(() => ({
-            logUserEvent: (linkData: TealiumEventObjectTypes) => {
+            logButtonEvent: () => {
                 return
             },
-            logPageView: () => {
-                return
-            },
-            logButtonEvent: (
-                tealiumData: Omit<TealiumButtonEventObject, 'event_name'>
-            ) => {
-                return
-            },
-            logInternalLinkEvent: (
-                tealiumData: Omit<TealiumInternalLinkEventObject, 'event_name'>
-            ) => {
+            logInternalLinkEvent: () => {
                 return
             },
         }))
