+++ services/app-web/src/tealium/index.ts
@@ -1,0 +1,20 @@
+export {
+    tealiumClient,
+    CONTENT_TYPE_BY_ROUTE,
+} from './tealium'
+
+export {
+    getTealiumPageName,
+    getTealiumEnv
+} from './tealiumHelpers'
+
+export type {
+    TealiumLinkDataObject,
+    TealiumViewDataObject,
+    TealiumEvent,
+    TealiumButtonEventObject,
+    TealiumInternalLinkEventObject,
+    TealiumEventObjectTypes,
+    TealiumClientType,
+    TealiumEnv
+} from './tealium'
