+++ services/app-web/src/testHelpers/tealiumHelpers.ts
@@ -1,7 +1,10 @@
-import { TealiumClientType } from '../constants/tealium';
+import { TealiumClientType } from '../tealium';
 
 export const tealiumTestClient = (): TealiumClientType => {
     return {
+        initializeTealium: () => {
+            return
+        },
         logUserEvent: () => {
             return
         },
